move SocketIOAMFlowClientOverride.ts to bare/playlogClient
