+++ packages/akashic-cli-serve/www/public/external/playlogClientV0_0_1.js
@@ -453,6 +453,18 @@ eval("\nmodule.exports = str => encodeURIComponent(str).replace(/[!'()*]/g, x =>
 
 /***/ }),
 
+/***/ "./src/bare/playlogClient/SocketIOAMFlowClientOverride.ts":
+/*!****************************************************************!*\
+  !*** ./src/bare/playlogClient/SocketIOAMFlowClientOverride.ts ***!
+  \****************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports, __webpack_require__) {
+
+"use strict";
+eval("\nvar __extends = (this && this.__extends) || (function () {\n    var extendStatics = function (d, b) {\n        extendStatics = Object.setPrototypeOf ||\n            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||\n            function (d, b) { for (var p in b) if (Object.prototype.hasOwnProperty.call(b, p)) d[p] = b[p]; };\n        return extendStatics(d, b);\n    };\n    return function (d, b) {\n        if (typeof b !== \"function\" && b !== null)\n            throw new TypeError(\"Class extends value \" + String(b) + \" is not a constructor or null\");\n        extendStatics(d, b);\n        function __() { this.constructor = d; }\n        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());\n    };\n})();\nObject.defineProperty(exports, \"__esModule\", { value: true });\nexports.SocketIOAMFlowClientOverride = void 0;\nvar SocketIOAMFlowClient_1 = __webpack_require__(/*! ../../client/akashic/SocketIOAMFlowClient */ \"./src/client/akashic/SocketIOAMFlowClient.ts\");\n/**\n * SocketIOAMFlowClientの関数を書き換えるためのクラス\n * 本来はSocketIOAMFlowClientの継承ではなく委譲にすべきだが、現状書き換えが必要なメソッドが少ないので継承とした。\n * TODO: 書き換えがメソッドが多くなったら委譲に切り替える。\n */\nvar SocketIOAMFlowClientOverride = /** @class */ (function (_super) {\n    __extends(SocketIOAMFlowClientOverride, _super);\n    function SocketIOAMFlowClientOverride(socket, params) {\n        var _this = _super.call(this, socket) || this;\n        _this._playId = params.playId;\n        _this._token = params.token;\n        return _this;\n    }\n    SocketIOAMFlowClientOverride.prototype.open = function (_playId, callback) {\n        _super.prototype.open.call(this, this._playId, callback);\n    };\n    SocketIOAMFlowClientOverride.prototype.authenticate = function (_token, callback) {\n        _super.prototype.authenticate.call(this, this._token, callback);\n    };\n    return SocketIOAMFlowClientOverride;\n}(SocketIOAMFlowClient_1.SocketIOAMFlowClient));\nexports.SocketIOAMFlowClientOverride = SocketIOAMFlowClientOverride;\n\n\n//# sourceURL=webpack://playlogClientV0_0_1/./src/bare/playlogClient/SocketIOAMFlowClientOverride.ts?");
+
+/***/ }),
+
 /***/ "./src/bare/playlogClient/playlogClientV0_0_1.ts":
 /*!*******************************************************!*\
   !*** ./src/bare/playlogClient/playlogClientV0_0_1.ts ***!
@@ -461,7 +473,7 @@ eval("\nmodule.exports = str => encodeURIComponent(str).replace(/[!'()*]/g, x =>
 /***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
-eval("\nObject.defineProperty(exports, \"__esModule\", { value: true });\nexports.Socket = exports.Session = void 0;\nvar SocketIOAMFlowClientOverride_1 = __webpack_require__(/*! ../../client/akashic/SocketIOAMFlowClientOverride */ \"./src/client/akashic/SocketIOAMFlowClientOverride.ts\");\nvar socketInstance_1 = __webpack_require__(/*! ../../client/api/socketInstance */ \"./src/client/api/socketInstance.ts\");\nvar ApiClient_1 = __webpack_require__(/*! ../../client/api/ApiClient */ \"./src/client/api/ApiClient.ts\");\nvar SocketIOAMFlowClient_1 = __webpack_require__(/*! ../../client/akashic/SocketIOAMFlowClient */ \"./src/client/akashic/SocketIOAMFlowClient.ts\");\n// playlogClientのplaylogServerUrlとして以下のURLを指定された時だけ動作確認テスト用に書き換えたAmflowを使う\nvar DUMMY_URLS = [\"http://dummy.playlog.net\"];\nvar DUMMY_PLAYER_ID = \"dummy\";\nvar Session = /** @class */ (function () {\n    function Session(url) {\n        this._url = url;\n    }\n    Session.prototype.open = function (cb) {\n        cb();\n    };\n    Session.prototype.on = function (_msg, _cb) { };\n    Session.prototype.createClient = function (opt, cb) {\n        if (typeof opt === \"function\") {\n            cb = opt;\n            opt = null;\n        }\n        var host = typeof HOST !== \"undefined\" ? HOST : window.location.host;\n        if (DUMMY_URLS.indexOf(this._url) !== -1) {\n            var protocol = window.location.protocol.includes(\"https\") ? \"https\" : \"http\";\n            var apiClient = new ApiClient_1.ApiClient(protocol + \"://\" + host);\n            // TODO: 外部からplayIdの指定をできるようにすべきだが、現状動作確認テストでは1つのplayしか使用しないので今の所はこのままでも問題ない\n            var playId_1 = \"0\";\n            apiClient.createPlayToken(playId_1, DUMMY_PLAYER_ID, false, DUMMY_PLAYER_ID, undefined).then(function (result) {\n                cb(null, new SocketIOAMFlowClientOverride_1.SocketIOAMFlowClientOverride((0, socketInstance_1.socketInstance)(host), { playId: playId_1, token: result.data.playToken }));\n            }).catch(function (err) { return cb(err); });\n        }\n        else {\n            cb(null, new SocketIOAMFlowClient_1.SocketIOAMFlowClient((0, socketInstance_1.socketInstance)()));\n        }\n    };\n    return Session;\n}());\nexports.Session = Session;\nexports.Socket = {\n    Type: {\n        WebSocket: 1\n    }\n};\n\n\n//# sourceURL=webpack://playlogClientV0_0_1/./src/bare/playlogClient/playlogClientV0_0_1.ts?");
+eval("\nObject.defineProperty(exports, \"__esModule\", { value: true });\nexports.Socket = exports.Session = void 0;\nvar socketInstance_1 = __webpack_require__(/*! ../../client/api/socketInstance */ \"./src/client/api/socketInstance.ts\");\nvar ApiClient_1 = __webpack_require__(/*! ../../client/api/ApiClient */ \"./src/client/api/ApiClient.ts\");\nvar SocketIOAMFlowClient_1 = __webpack_require__(/*! ../../client/akashic/SocketIOAMFlowClient */ \"./src/client/akashic/SocketIOAMFlowClient.ts\");\nvar SocketIOAMFlowClientOverride_1 = __webpack_require__(/*! ./SocketIOAMFlowClientOverride */ \"./src/bare/playlogClient/SocketIOAMFlowClientOverride.ts\");\n// playlogClientのplaylogServerUrlとして以下のURLを指定された時だけ動作確認テスト用に書き換えたAmflowを使う\nvar DUMMY_URLS = [\"http://dummy.playlog.net\"];\nvar DUMMY_PLAYER_ID = \"dummy\";\nvar Session = /** @class */ (function () {\n    function Session(url) {\n        this._url = url;\n    }\n    Session.prototype.open = function (cb) {\n        cb();\n    };\n    Session.prototype.on = function (_msg, _cb) { };\n    Session.prototype.createClient = function (opt, cb) {\n        if (typeof opt === \"function\") {\n            cb = opt;\n            opt = null;\n        }\n        var host = typeof HOST !== \"undefined\" ? HOST : window.location.host;\n        if (DUMMY_URLS.indexOf(this._url) !== -1) {\n            var protocol = window.location.protocol.includes(\"https\") ? \"https\" : \"http\";\n            var apiClient = new ApiClient_1.ApiClient(protocol + \"://\" + host);\n            // TODO: 外部からplayIdの指定をできるようにすべきだが、現状動作確認テストでは1つのplayしか使用しないので今の所はこのままでも問題ない\n            var playId_1 = \"0\";\n            apiClient.createPlayToken(playId_1, DUMMY_PLAYER_ID, false, DUMMY_PLAYER_ID, undefined).then(function (result) {\n                cb(null, new SocketIOAMFlowClientOverride_1.SocketIOAMFlowClientOverride((0, socketInstance_1.socketInstance)(host), { playId: playId_1, token: result.data.playToken }));\n            }).catch(function (err) { return cb(err); });\n        }\n        else {\n            cb(null, new SocketIOAMFlowClient_1.SocketIOAMFlowClient((0, socketInstance_1.socketInstance)()));\n        }\n    };\n    return Session;\n}());\nexports.Session = Session;\nexports.Socket = {\n    Type: {\n        WebSocket: 1\n    }\n};\n\n\n//# sourceURL=webpack://playlogClientV0_0_1/./src/bare/playlogClient/playlogClientV0_0_1.ts?");
 
 /***/ }),
 
@@ -474,18 +486,6 @@ eval("\nObject.defineProperty(exports, \"__esModule\", { value: true });\nexport
 
 "use strict";
 eval("/* WEBPACK VAR INJECTION */(function(Promise) {\nObject.defineProperty(exports, \"__esModule\", { value: true });\nexports.SocketIOAMFlowClient = void 0;\nvar trigger_1 = __webpack_require__(/*! @akashic/trigger */ \"./node_modules/@akashic/trigger/lib/index.js\");\nvar SocketIOAMFlowClient = /** @class */ (function () {\n    function SocketIOAMFlowClient(socket) {\n        var _this = this;\n        this._onTick = function (connectionId, tick) {\n            if (connectionId !== _this._connectionId)\n                return;\n            _this._tickHandlers.forEach(function (h) { return h(tick); });\n        };\n        this._onEvent = function (connectionId, pev) {\n            if (connectionId !== _this._connectionId)\n                return;\n            _this._eventHandlers.forEach(function (h) { return h(pev); });\n        };\n        this.onGotStartedAt = new trigger_1.Trigger();\n        this._socket = socket;\n        this._tickHandlers = [];\n        this._eventHandlers = [];\n        this._permission = null;\n        this._startedAt = null;\n        this._eventQueue = [];\n        this._connectionId = null;\n    }\n    /**\n     * 現状開いているplayの開始時刻を取得する。\n     * SocketIOAMFlowClient独自拡張。暫定。\n     *\n     * この値が null の場合は、以降 (最初に getStartPoint() が呼ばれた時点で) onGotStartedAt がfireされる。\n     *\n     * TODO 開始時刻は、おそらくサーバ側で putStartPoint() に反応してpush通知する方がクリーン。\n     * その場合 getPlay() の情報にも開始時刻を含めることになる。\n     * ただし現状では Runner からの putStartPoint() を監視する方法がない。\n     */\n    SocketIOAMFlowClient.prototype.getStartedAt = function () {\n        return this._startedAt;\n    };\n    /**\n     * 可能なタイミングでイベントを送る。\n     * SocketIOAMFlowClient独自拡張。暫定。\n     *\n     * TODO playIdでイベントを投げるAPIを作り、そちらを利用する。\n     */\n    SocketIOAMFlowClient.prototype.enqueueEvent = function (event) {\n        if (this._permission != null) { // 暫定でなければpermissionの内容を踏まえるべき\n            this.sendEvent(event);\n        }\n        else {\n            this._eventQueue.push(event);\n        }\n    };\n    /**\n     * getStartPoint() の Promise 版。\n     * SocketIOAMFlowClient独自拡張。\n     */\n    SocketIOAMFlowClient.prototype.getStartPointPromise = function (opts) {\n        var _this = this;\n        return new Promise(function (resolve, reject) {\n            _this.getStartPoint(opts, function (err, sp) {\n                void (err ? reject(err) : resolve(sp));\n            });\n        });\n    };\n    SocketIOAMFlowClient.prototype.open = function (playId, callback) {\n        var _this = this;\n        this._socket.on(\"amflow:[tick]\", this._onTick);\n        this._socket.on(\"amflow:[event]\", this._onEvent);\n        this._socket.emit(\"amflow:open\", playId, function (err, connectionId) {\n            if (err) {\n                callback(err);\n                return;\n            }\n            _this._connectionId = connectionId;\n            callback();\n        });\n    };\n    SocketIOAMFlowClient.prototype.close = function (callback) {\n        this._socket.off(\"amflow:[tick]\", this._onTick);\n        this._socket.off(\"amflow:[event]\", this._onEvent);\n        this._socket.emit(\"amflow:close\", this._connectionId, callback);\n        this._connectionId = null;\n    };\n    SocketIOAMFlowClient.prototype.authenticate = function (token, callback) {\n        var _this = this;\n        this._socket.emit(\"amflow:authenticate\", this._connectionId, token, function (error, permission) {\n            _this._permission = permission;\n            _this._eventQueue.forEach(function (ev) { return _this.sendEvent(ev); });\n            _this._eventQueue = [];\n            callback(error, permission);\n        });\n    };\n    SocketIOAMFlowClient.prototype.sendTick = function (tick) {\n        this._socket.emit(\"amflow:sendTick\", this._connectionId, tick);\n    };\n    SocketIOAMFlowClient.prototype.onTick = function (handler) {\n        this._tickHandlers.push(handler);\n        if (this._tickHandlers.length === 1) {\n            this._socket.emit(\"amflow:onTick\", this._connectionId);\n        }\n    };\n    SocketIOAMFlowClient.prototype.offTick = function (handler) {\n        this._tickHandlers = this._tickHandlers.filter(function (h) { return h !== handler; });\n        if (this._tickHandlers.length === 0) {\n            this._socket.emit(\"amflow:offTick\", this._connectionId);\n        }\n    };\n    SocketIOAMFlowClient.prototype.sendEvent = function (event) {\n        this._socket.emit(\"amflow:sendEvent\", this._connectionId, event);\n    };\n    SocketIOAMFlowClient.prototype.onEvent = function (handler) {\n        this._eventHandlers.push(handler);\n        if (this._eventHandlers.length === 1) {\n            this._socket.emit(\"amflow:onEvent\", this._connectionId);\n        }\n    };\n    SocketIOAMFlowClient.prototype.offEvent = function (handler) {\n        this._eventHandlers = this._eventHandlers.filter(function (h) { return h !== handler; });\n        if (this._eventHandlers.length === 0) {\n            this._socket.emit(\"amflow:offEvent\", this._connectionId);\n        }\n    };\n    SocketIOAMFlowClient.prototype.getTickList = function (optsOrBegin, endOrCallback, callbackOrUndefined) {\n        var opts;\n        var callback;\n        if (typeof optsOrBegin === \"number\") {\n            // NOTE: optsOrBegin === \"number\" であれば必ず amflow@2 以前の引数だとみなしてキャストする\n            opts = {\n                begin: optsOrBegin,\n                end: endOrCallback\n            };\n            callback = callbackOrUndefined;\n        }\n        else {\n            // NOTE: optsOrBegin !== \"number\" であれば必ず amflow@3 以降の引数だとみなしてキャストする\n            opts = optsOrBegin;\n            callback = endOrCallback;\n        }\n        this._socket.emit(\"amflow:getTickList\", this._connectionId, opts, callback);\n    };\n    SocketIOAMFlowClient.prototype.putStartPoint = function (startPoint, callback) {\n        this._socket.emit(\"amflow:putStartPoint\", this._connectionId, startPoint, callback);\n    };\n    SocketIOAMFlowClient.prototype.getStartPoint = function (opts, callback) {\n        var _this = this;\n        this._socket.emit(\"amflow:getStartPoint\", this._connectionId, opts, function (err, startPoint) {\n            if ((_this._startedAt === null) && !err && opts.frame === 0) {\n                _this._startedAt = startPoint.data.startedAt;\n                _this.onGotStartedAt.fire(_this._startedAt);\n            }\n            callback(err, startPoint);\n        });\n    };\n    // 継承元のinterfaceに合わせるためのメソッド定義。noUnusedParametersを無視するために利用されていない引数のprefixに_を付けた\n    SocketIOAMFlowClient.prototype.putStorageData = function (_key, _value, _options, callback) {\n        callback(new Error(\"not supported\"));\n    };\n    // 継承元のinterfaceに合わせるためのメソッド定義。noUnusedParametersを無視するために利用されていない引数のprefixに_を付けた\n    SocketIOAMFlowClient.prototype.getStorageData = function (_keys, callback) {\n        callback(new Error(\"not supported\"), null);\n    };\n    return SocketIOAMFlowClient;\n}());\nexports.SocketIOAMFlowClient = SocketIOAMFlowClient;\n\n/* WEBPACK VAR INJECTION */}.call(this, __webpack_require__(/*! es6-promise */ \"./node_modules/es6-promise/dist/es6-promise.js\")))\n\n//# sourceURL=webpack://playlogClientV0_0_1/./src/client/akashic/SocketIOAMFlowClient.ts?");
-
-/***/ }),
-
-/***/ "./src/client/akashic/SocketIOAMFlowClientOverride.ts":
-/*!************************************************************!*\
-  !*** ./src/client/akashic/SocketIOAMFlowClientOverride.ts ***!
-  \************************************************************/
-/*! no static exports found */
-/***/ (function(module, exports, __webpack_require__) {
-
-"use strict";
-eval("\nvar __extends = (this && this.__extends) || (function () {\n    var extendStatics = function (d, b) {\n        extendStatics = Object.setPrototypeOf ||\n            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||\n            function (d, b) { for (var p in b) if (Object.prototype.hasOwnProperty.call(b, p)) d[p] = b[p]; };\n        return extendStatics(d, b);\n    };\n    return function (d, b) {\n        if (typeof b !== \"function\" && b !== null)\n            throw new TypeError(\"Class extends value \" + String(b) + \" is not a constructor or null\");\n        extendStatics(d, b);\n        function __() { this.constructor = d; }\n        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());\n    };\n})();\nObject.defineProperty(exports, \"__esModule\", { value: true });\nexports.SocketIOAMFlowClientOverride = void 0;\nvar SocketIOAMFlowClient_1 = __webpack_require__(/*! ./SocketIOAMFlowClient */ \"./src/client/akashic/SocketIOAMFlowClient.ts\");\n/**\n * SocketIOAMFlowClientの関数を書き換えるためのクラス\n * 本来はSocketIOAMFlowClientの継承ではなく委譲にすべきだが、現状書き換えが必要なメソッドが少ないので継承とした。\n * TODO: 書き換えがメソッドが多くなったら委譲に切り替える。\n */\nvar SocketIOAMFlowClientOverride = /** @class */ (function (_super) {\n    __extends(SocketIOAMFlowClientOverride, _super);\n    function SocketIOAMFlowClientOverride(socket, params) {\n        var _this = _super.call(this, socket) || this;\n        _this._playId = params.playId;\n        _this._token = params.token;\n        return _this;\n    }\n    SocketIOAMFlowClientOverride.prototype.open = function (_playId, callback) {\n        _super.prototype.open.call(this, this._playId, callback);\n    };\n    SocketIOAMFlowClientOverride.prototype.authenticate = function (_token, callback) {\n        _super.prototype.authenticate.call(this, this._token, callback);\n    };\n    return SocketIOAMFlowClientOverride;\n}(SocketIOAMFlowClient_1.SocketIOAMFlowClient));\nexports.SocketIOAMFlowClientOverride = SocketIOAMFlowClientOverride;\n\n\n//# sourceURL=webpack://playlogClientV0_0_1/./src/client/akashic/SocketIOAMFlowClientOverride.ts?");
 
 /***/ }),
